Move 2024 MerchCard to its own file in components
